chore: rename 1_1prime to 1_1
