+++ message/message.go
@@ -1,7 +1,7 @@
 package message
 
 import (
-	message1_1 "github.com/filecoin-project/go-data-transfer/message/message1_1prime"
+	"github.com/filecoin-project/go-data-transfer/message/message1_1"
 )
 
 var NewRequest = message1_1.NewRequest
