+++ message/message1_1/message_test.go
@@ -18,7 +18,7 @@ import (
 
 	datatransfer "github.com/filecoin-project/go-data-transfer"
 	"github.com/filecoin-project/go-data-transfer/ipldutil"
-	message1_1 "github.com/filecoin-project/go-data-transfer/message/message1_1prime"
+	message1_1 "github.com/filecoin-project/go-data-transfer/message/message1_1"
 	"github.com/filecoin-project/go-data-transfer/testutil"
 )
